requirements.txt
streamlit
openai
langchain-openai
langchain-core
python-dotenv
langchain_community
langchain
